Refactor PullRequestUpdater classes (#6154)
