--- src/ProductConstructionService/ProductConstructionService.DependencyFlow/BatchedDependencyPullRequestUpdater.cs
@@ -1,52 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-using Maestro.Common.Cache;
-using Maestro.Data;
-using Maestro.Data.Models;
-using Maestro.DataProviders;
-using Microsoft.DotNet.DarcLib;
-using Microsoft.Extensions.Logging;
-using ProductConstructionService.DependencyFlow.Model;
-using ProductConstructionService.WorkItems;
-
-namespace ProductConstructionService.DependencyFlow;
-
-internal class BatchedDependencyPullRequestUpdater : DependencyPullRequestUpdater
-{
-    private readonly BatchedPullRequestUpdaterId _id;
-    private readonly BuildAssetRegistryContext _context;
-
-    public BatchedDependencyPullRequestUpdater(
-        BatchedPullRequestUpdaterId id,
-        IMergePolicyEvaluator mergePolicyEvaluator,
-        BuildAssetRegistryContext context,
-        IRemoteFactory remoteFactory,
-        IPullRequestUpdaterFactory updaterFactory,
-        ICoherencyUpdateResolver coherencyUpdateResolver,
-        IPullRequestBuilder pullRequestBuilder,
-        IRedisCacheFactory cacheFactory,
-        IReminderManagerFactory reminderManagerFactory,
-        ISqlBarClient sqlClient,
-        ILogger<BatchedDependencyPullRequestUpdater> logger,
-        IPullRequestCommenter pullRequestCommenter)
-        : base(id, mergePolicyEvaluator, context, remoteFactory, updaterFactory, coherencyUpdateResolver, pullRequestBuilder, cacheFactory, reminderManagerFactory, sqlClient, logger, pullRequestCommenter)
-    {
-        _id = id;
-        _context = context;
-    }
-
-    protected override Task<(string repository, string branch)> GetTargetAsync()
-    {
-        return Task.FromResult((_id.Repository, _id.Branch));
-    }
-
-    protected override async Task<IReadOnlyList<MergePolicyDefinition>> GetMergePolicyDefinitions()
-    {
-        RepositoryBranch? repositoryBranch = await _context.RepositoryBranches.FindAsync(_id.Repository, _id.Branch);
-        return repositoryBranch?.PolicyObject?.MergePolicies ?? [];
-    }
-
-    // For batced subscriptions we don't know which source repo to tag
-    protected override Task TagSourceRepositoryGitHubContactsIfPossibleAsync(InProgressPullRequest pr) => Task.CompletedTask;
-}
